+++ hrp/cmd/convert.go
@@ -2,10 +2,14 @@ package cmd
 
 import (
 	"errors"
+	"fmt"
 
+	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 
+	"github.com/httprunner/httprunner/v4/hrp/internal/builtin"
 	"github.com/httprunner/httprunner/v4/hrp/internal/convert"
+	"github.com/httprunner/httprunner/v4/hrp/internal/version"
 )
 
 var convertCmd = &cobra.Command{
@@ -32,6 +36,15 @@ var convertCmd = &cobra.Command{
 		if toPyTestFlag {
 			flagCount++
 			outputType = convert.OutputTypePyTest
+
+			packages := []string{
+				fmt.Sprintf("httprunner==%s", version.VERSION),
+			}
+			_, err := builtin.EnsurePython3Venv(venv, packages...)
+			if err != nil {
+				log.Error().Err(err).Msg("python3 venv is not ready")
+				return err
+			}
 		}
 		if flagCount > 1 {
 			return errors.New("please specify at most one conversion flag")