+++ hrp/internal/scaffold/examples_test.go
@@ -1,30 +1,32 @@
 package scaffold
 
 import (
+	"path/filepath"
 	"testing"
 )
 
 func TestGenDemoExamples(t *testing.T) {
 	dir := "../../../examples/demo-with-go-plugin"
-	err := CreateScaffold(dir, Go, true)
+	err := CreateScaffold(dir, Go, "", true)
 	if err != nil {
 		t.Fatal()
 	}
 
 	dir = "../../../examples/demo-with-py-plugin"
-	err = CreateScaffold(dir, Py, true)
+	venv := filepath.Join(dir, ".venv")
+	err = CreateScaffold(dir, Py, venv, true)
 	if err != nil {
 		t.Fatal()
 	}
 
 	dir = "../../../examples/demo-without-plugin"
-	err = CreateScaffold(dir, Ignore, true)
+	err = CreateScaffold(dir, Ignore, "", true)
 	if err != nil {
 		t.Fatal()
 	}
 
 	dir = "../../../examples/demo-empty-project"
-	err = CreateScaffold(dir, Empty, true)
+	err = CreateScaffold(dir, Empty, "", true)
 	if err != nil {
 		t.Fatal()
 	}